# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stream/start:
    post:
      operationId: startStream
      tags: [streaming]
      summary: Start a live stream
      description: Starts FFmpeg and sends a selected local category video to the supplied ingest URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStartInput"
      responses:
        "202":
          description: Stream process accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamControlResponse"
        "400":
          description: Invalid stream configuration
        "409":
          description: Stream is already running
        "500":
          description: Stream process could not start
  /stream/stop:
    post:
      operationId: stopStream
      tags: [streaming]
      summary: Stop a live stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStopInput"
      responses:
        "200":
          description: Stream stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamControlResponse"
        "404":
          description: Stream is not running
  /stream/update:
    post:
      operationId: updateStream
      tags: [streaming]
      summary: Update the playlist of a running stream
      description: Rebuilds the active FFmpeg playlist so videos added to or removed from a category take effect while streaming.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStartInput"
      responses:
        "202":
          description: Stream playlist update accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamControlResponse"
        "400":
          description: Invalid stream playlist
        "404":
          description: Stream is not running
  /stream/status/{streamId}:
    get:
      operationId: getStreamStatus
      tags: [streaming]
      summary: Get stream process status
      parameters:
        - name: streamId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stream status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamControlResponse"
...
components:
...
  schemas:
...
    StreamStartInput:
      type: object
      properties:
        streamId:
          type: string
          minLength: 1
        ingestUrl:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        videoSource:
          type: string
        videoSources:
          type: array
          items:
            type: string
            minLength: 1
        faceCategory:
          type: string
        faceSource:
          type: string
        faceSources:
          type: array
          items:
            type: string
        playbackSpeed:
          type: number
          minimum: 0.5
          maximum: 2
          default: 1
        quality:
          type: string
          enum: [4k, 1080p]
          default: 1080p
        aspectRatio:
          type: string
          enum: [shorts, full, square]
          default: full
        facePosition:
          type: string
          enum: [top-left, top-right, bottom-left, bottom-right, center]
          default: bottom-right
        faceScale:
          type: number
          minimum: 0.1
          maximum: 0.6
          default: 0.25
        durationMinutes:
          type: number
          minimum: 1
          maximum: 1440
        autoRestart:
          type: boolean
          default: false
        voiceAudio:
          type: boolean
          default: false
          description: Keeps a server-side microphone PCM input ready for live voice-over.
        liveAnimationSource:
          type: string
          description: Optional server-ready animation video to layer above the live playlist.
        liveAnimationX:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        liveAnimationY:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        liveAnimationScale:
          type: number
          minimum: 0.1
          maximum: 0.8
          default: 0.25
        composition:
          $ref: '#/components/schemas/StreamCompositionInput'
      required:
        - streamId
        - ingestUrl
        - category
    StreamCompositionInput:
      type: object
      description: Optional editor composition to apply directly while the live stream is encoded.
      properties:
        mainX:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        mainY:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        mainScale:
          type: number
          minimum: 0.5
          maximum: 2.5
          default: 1
        cropMode:
          type: string
          enum: [fit, crop]
          default: fit
        webcamSource:
          type: string
        webcamX:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        webcamY:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        webcamScale:
          type: number
          minimum: 0.1
          maximum: 0.8
          default: 0.25
        animationSource:
          type: string
        animationX:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        animationY:
          type: number
          minimum: -48
          maximum: 48
          default: 0
        animationScale:
          type: number
          minimum: 0.1
          maximum: 0.8
          default: 0.25
        animationLayers:
          type: array
          description: Multiple animation overlays that play together in the editor composition.
          items:
            type: object
            required: [source]
            properties:
              source:
                type: string
              x:
                type: number
                minimum: -48
                maximum: 48
                default: 0
              y:
                type: number
                minimum: -48
                maximum: 48
                default: 0
              scale:
                type: number
                minimum: 0.1
                maximum: 0.8
                default: 0.25
        logoSource:
          type: string
        logoPosition:
          type: string
          enum: [top-left, top-right, bottom-left, bottom-right]
          default: bottom-right
        logoScale:
          type: number
          minimum: 0.1
          maximum: 0.6
          default: 0.25
        animationPreset:
          type: string
          enum: [none, subscribe, like, follow]
          default: none
        comingSoon:
          type: boolean
          default: false
          description: Show a COMING SOON lower-third in the live composition.
        brightness:
          type: number
          minimum: -1
          maximum: 1
          default: 0
        contrast:
          type: number
          minimum: 0.5
          maximum: 1.8
          default: 1
        saturation:
          type: number
          minimum: 0
          maximum: 2
          default: 1
        hue:
          type: number
          minimum: -180
          maximum: 180
          default: 0
        chromaKeyBySource:
          type: object
          additionalProperties:
            type: object
            properties:
              enabled:
                type: boolean
              color:
                type: string
                pattern: "^#?[0-9a-fA-F]{6}$"
              similarity:
                type: number
                minimum: 0.05
                maximum: 0.95
              blend:
                type: number
                minimum: 0
                maximum: 0.5
        chromaKeyDurations:
          type: object
          additionalProperties:
            type: number
            minimum: 0
        chromaKeyByLayer:
          type: object
          additionalProperties:
            type: object
            properties:
              enabled:
                type: boolean
              color:
                type: string
                pattern: "^#?[0-9a-fA-F]{6}$"
              similarity:
                type: number
                minimum: 0.05
                maximum: 0.95
              blend:
                type: number
                minimum: 0
                maximum: 0.5
        chromaKeyEnabled:
          type: boolean
          default: false
        chromaKeyTarget:
          type: string
          enum: [main, webcam, animation]
          default: webcam
        chromaKeyColor:
          type: string
          default: '#00ff00'
        chromaSimilarity:
          type: number
          minimum: 0.1
          maximum: 0.9
          default: 0.32
        chromaBlend:
          type: number
          minimum: 0
          maximum: 0.35
          default: 0.08
    StreamStopInput:
      type: object
      properties:
        streamId:
          type: string
          minLength: 1
      required:
        - streamId
    StreamControlResponse:
      type: object
      properties:
        streamId:
          type: string
        status:
          type: string
          enum: [running, stopped, failed]
        message:
          type: string
        pid:
          type: ["number", "null"]
      required:
        - streamId
        - status
        - message
        - pid
...